리뷰리스트 생성 기능 구현

+++ app/src/main/java/com/hjh/likecafe/ReviewVO.java
@@ -6,11 +6,11 @@ public class ReviewVO {
     private int cafe_id;
     private String cafe_name;
     private String mem_id;
-    private int star;
+    private double star;
     private String content;
     private int image;
 
-    public ReviewVO(int id, int cafe_id, String cafe_name, String mem_id, int star, String content, int image) {
+    public ReviewVO(int id, int cafe_id, String cafe_name, String mem_id, double star, String content, int image) {
         this.id = id;
         this.cafe_id = cafe_id;
         this.cafe_name = cafe_name;
@@ -36,7 +36,7 @@ public class ReviewVO {
         return mem_id;
     }
 
-    public int getStar() {
+    public double getStar() {
         return star;
     }
